Reorganize trade folder
