+++ src/components/mining/TokenInput.tsx
