+++ src/components/trade/_shared/QuickTradeSelector.tsx
