+++ src/components/trade/_shared/QuickTradeSettingsPopover.tsx
