+++ src/components/trade/_shared/footer/ContractExecutionView.tsx
