+++ src/components/trade/_shared/footer/ProtectionWarning.tsx
