+++ src/components/trade/index.tsx
@@ -8,9 +8,9 @@ import { useNetwork } from 'hooks/useNetwork'
 import { SlippageProvider, useSlippage } from 'providers/Slippage'
 import { isTokenAvailableForFlashMint } from 'utils/tokens'
 
-import FlashMint from './FlashMint'
-import QuickTrade, { QuickTradeProps } from './QuickTrade'
-import { QuickTradeSettingsPopover } from './QuickTradeSettingsPopover'
+import { QuickTradeSettingsPopover } from './_shared/QuickTradeSettingsPopover'
+import FlashMint from './flashmint'
+import QuickTrade, { QuickTradeProps } from './swap'
 
 enum TradeType {
   flashMint,
